refactor: move trainers to different files
